requirements.txt
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
cssselect2==0.3.0
decorator==4.4.2
defusedxml==0.6.0
future==0.18.2
idna==2.8
lxml==4.4.2
numpy==1.19.1
Pillow==7.2.0
pycparser==2.20
python-telegram-bot==12.2.0
reportlab==3.5.46
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
svglib==1.0.0
tinycss2==1.0.2
tornado==6.0.4
urllib3==1.25.10
webencodings==0.5.1
